This code much clenaer
